devided the packages to main, fortune, fortune_test
