+++ kadai4/Mizushima/.gitignore
@@ -1,1 +1,2 @@
 bin
+profile
