+++ kadai4/Mizushima/fortune/export_test.go
@@ -1,0 +1,7 @@
+package fortune
+
+import "time"
+
+func SetTime(t time.Time) {
+	timeNow = func() time.Time { return t }
+}
